refactor(kernel): split the cross-operand plane weld out of mesh_bridge
`promote_cutter_verts_onto_host_faces` and its `exact_on_plane_weld` helper
move verbatim to `kernel/plane_weld.rs`. They were subtraction-only helpers
living in `mesh_bridge.rs`, which sat at its module-size budget; #3353 makes
the weld a boolean-wide concern, so it gets its own home before it grows a
second caller.

Pure move: no logic changes. `mesh_bridge.rs` drops from 484 to 298 lines and
loses its allowlist row (it is now under the 400-line limit), so the pinned
rust/geometry digest moves with it.

Refs #3353
