+++ rust/geometry/src/kernel/mod.rs
@@ -29,6 +29,7 @@ pub mod manifest;
 pub mod mesh_bridge;
 pub mod mesh_volume;
 pub(crate) mod near_band;
+pub(crate) mod plane_weld;
 pub mod predicates;
 pub mod rational;
 pub(crate) mod signed_volume;
